refactor(storage): share the Lore client via @yucp/shared
