+++ apps/api/src/index.ts
@@ -4,6 +4,7 @@
 
 import path from 'node:path';
 import { getInternalRpcSharedSecret } from '@yucp/shared';
+import type { LoreBackstageConfig } from '@yucp/shared/loreBackstageClient';
 import { type Auth, createAuth } from './auth';
 import { createInternalRpcRouter, INTERNAL_RPC_PATH } from './internalRpc/router';
 import { getClientAddress } from './lib/clientAddress';
@@ -18,7 +19,6 @@ import {
   isLegacyFrontendAsset,
 } from './lib/legacyFrontend';
 import { logger } from './lib/logger';
-import type { LoreBackstageConfig } from './lib/loreBackstage';
 import {
   annotateApiSpan,
   getActiveTraceIds,
