+++ packages/shared/package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "@infisical/sdk": "^4.0.6",
+    "@noble/hashes": "^2.0.1",
     "@opentelemetry/api": "^1.9.1",
     "@opentelemetry/context-async-hooks": "^2.8.0",
     "@opentelemetry/core": "^2.8.0",
